Refactor security package
